requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
python-dotenv
apscheduler